requirements.txt
requests
python-dotenv
pytest
mysql-connector-python
